content/media/test/test_audio2.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: Audio Constructor Test 2</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">
var tmpAudio = new Audio();
for (var i = 0; i < gAudioTests.length; ++i) {
  var test = gAudioTests[i];
  if (!tmpAudio.canPlayType(test.type))
    continue;

  var a1 = new Audio(test.name);
  is(a1.getAttribute("autobuffer"), "autobuffer", "Autobuffer automatically set");
  is(a1.src.match("/[^/]+$"), "/" + test.name, "src OK");
}
</script>
</pre>
</body>
</html>
